+++ packages/ui/src/components/Text/index.tsx
@@ -1,7 +1,12 @@
-export const TextBody1 = ({ children }: { children: React.ReactNode }): JSX.Element => {
-  return <div className="text-gray-500 text-body1">{children}</div>;
+type TextProps = {
+  children: React.ReactNode;
+  className?: string;
 };
 
-export const TextBody2 = ({ children }: { children: React.ReactNode }): JSX.Element => {
-  return <div className="text-gray-500 text-body2">{children}</div>;
+export const TextBody1 = ({ children, className }: TextProps): JSX.Element => {
+  return <div className={`text-body1 text-gray-500 ${className}`}>{children}</div>;
+};
+
+export const TextBody2 = ({ children, className }: TextProps): JSX.Element => {
+  return <div className={`text-body2 text-gray-500 ${className}`}>{children}</div>;
 };
